完成登陆验证

+++ src/main/java/cn/edu/upc/yb/integrate/homepage/controller/TabController.java
@@ -1,6 +1,7 @@
 package cn.edu.upc.yb.integrate.homepage.controller;
 
 import cn.edu.upc.yb.integrate.calendar.dto.JsonMes;
+import cn.edu.upc.yb.integrate.common.service.CommonAdminService;
 import cn.edu.upc.yb.integrate.homepage.model.Tab;
 import cn.edu.upc.yb.integrate.homepage.repository.TabRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,30 +18,50 @@ public class TabController {
     @Autowired
     TabRepository tabRepository;
 
+    @Autowired
+    CommonAdminService commonAdminService;
+
     @GetMapping("/showall")
     public Object showAll(){
-        return tabRepository.findAll();
+        if(commonAdminService.isCommonAdmin()){
+            return tabRepository.findAll();
+        }else{
+            return new JsonMes(1,"你还没有登陆或者你不是管理员");
+        }
     }
 
     @GetMapping("/create")
     public Object create(String name, String href){
-        Tab tab = new Tab(name,href);
-        tabRepository.save(tab);
-        return new JsonMes(0,"保存成功");
+        if(commonAdminService.isCommonAdmin()){
+            Tab tab = new Tab(name,href);
+            tabRepository.save(tab);
+            return new JsonMes(0,"保存成功");
+        }else{
+            return new JsonMes(1,"你还没有登陆或者你不是管理员");
+        }
+
     }
 
     @GetMapping("/update")
     public Object update(Integer id, String name, String href){
-        Tab tab = tabRepository.findOne(id);
-        tab.update(name,href);
-        tabRepository.save(tab);
-        return new JsonMes(0,"更新成功");
+        if(commonAdminService.isCommonAdmin()){
+            Tab tab = tabRepository.findOne(id);
+            tab.update(name,href);
+            tabRepository.save(tab);
+            return new JsonMes(0,"更新成功");
+        }else{
+            return new JsonMes(1,"你还没有登陆或者你不是管理员");
+        }
     }
 
     @GetMapping("/delete")
     public Object delete(Integer id){
-        tabRepository.delete(id);
-        return new JsonMes(0,"删除成功");
+        if(commonAdminService.isCommonAdmin()){
+            tabRepository.delete(id);
+            return new JsonMes(0,"删除成功");
+        }else{
+            return new JsonMes(1,"你还没有登陆或者你不是管理员");
+        }
     }
